DRIVERS/VME_16Z002: remove potential deadlocks from vme4l-pldz002-cham.c
Any function with a PLDZ002_(UN)LOCK_STATE (except PldZ002Irq) can be
interrupted with IRQ when in critical section. In that case PldZ002Irq will
wait forever on a spinlock. It was never observed in practice.

It is probably possible to replace some of these spinlocks turning off
IRQ with another spinlock which will be not used in the IRQ handler.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>