Fix MDIS building Makefiles.
kernelsettings Makefile was causing a bad behaviour due to the include
of Kernel's headers Makefile. The Kernel's headers Makefile should never
be invoked in the context of the MDIS kernel modules compilation.

Now the compiler and linker are not retrieved from Kernel's header Makefile
but explicitly named in the kernelsettings Makefile.

Also add support for effective MDIS cross-compilation.

This commit fixes bugs #232 and #246