INSTALL.sh: Use tag date instead of last commit date in STAMPED_VERSION
Currently, the STAMPED_VERSION that is printed in each submodule's Makefile
is got from the latest commit date, however, if more than one tag point
to the same commit hash, the STAMPED_VERSION will be always the same.

The changes proposed here just get the tag date so every time a new
tag is created, the STAMPED_VERSION will change, even if it points
to the same commit.